requirements.txt
blinker==1.9.0
click==8.1.8
colorama==0.4.6
docx2pdf==0.1.8
fire==0.7.0
Flask==3.1.0
fonttools==4.57.0
gunicorn==23.0.0
itsdangerous==2.2.0
Jinja2==3.1.6
lxml==5.3.2
MarkupSafe==3.0.2
numpy==2.2.4
opencv-python-headless==4.11.0.86
packaging==24.2
pdf2docx==0.5.8
PyMuPDF==1.25.5
python-docx==1.1.2
python-dotenv==1.1.0
termcolor==3.0.1
tqdm==4.67.1
typing_extensions==4.13.2
Werkzeug==3.1.3
python-docx2pdf==0.1.7
reportlab==3.6.12